demo-collab/requirements.txt
flask>=3.0
PyJWT>=2.0
